make feed Intent Filters targetted to only feeds
The original IntentFilters made Courier show up as an option for all http
and https links.  But Courier cannot handle all http/s links.  So for some
random webpage, the link is placed in the "Add Feed" field, but when you
click Add, it fails unless its an RSS feed.

This also makes clicking on a proper feed work better since Courier provides
the only matching pattern, when the pattern is more specific, and therefore
Courier is the option option, and the process skips the menu to chose which
app, and instead the feed goes straight into the Add Feed page.

closes #4452 https://dev.guardianproject.info/issues/4452